Use custom render method to add Providers to App following doc
